feat: implement TradeScreen component with WebSocket integration and dynamic player updates
